Trending-api merge problem corrected
For some reason, the merge between profile and trending-api branchs created some problems: some files that were modified wasn't showing differences and no matter what was done, some files were merging even when told to not to. So a manually merge was done. But in this process, some code was lost. But the most important systems were saved.