Tag system revised (still only working on videos)
(Only working on video submission) The insert entry system was changed to a new one based on search. The tag system was revised to make it not dependent of videos. The navbar entry search was modifed to not conflict with other searchs systems. The modelform of the videos and images were changed to not have tags or entries (avoid load the entire database that would not be used) and these fields are set into de models 'manually'.