Origin Date reimplemented
The origin date field was reimplemented to use selection boxes, that will garantee a clear pattern in date storage.
It's still needed to verify the vality of the image inputs as well the tags in the view. It's needed too prevent the user to create and entry with a already used name.